style: break up styles per custom element
