+++ demo/index.html
@@ -27,20 +27,20 @@
       <demo-snippet>
         <template>
           <div style="width: 528px">
-            <auro-alaska customSize alaska></auro-alaska>
-            <auro-alaska customSize alaskaTagline></auro-alaska>
+            <auro-alaska alaska></auro-alaska>
+            <auro-alaska alaskaTagline></auro-alaska>
           </div>
           <div style="width: 192px">
-            <auro-alaska customSize alaska></auro-alaska>
-            <auro-alaska customSize alaskaTagline></auro-alaska>
+            <auro-alaska alaska></auro-alaska>
+            <auro-alaska alaskaTagline></auro-alaska>
           </div>
           <div style="width: 108px">
-            <auro-alaska customSize alaska></auro-alaska>
-            <auro-alaska customSize alaskaTagline></auro-alaska>
+            <auro-alaska alaska></auro-alaska>
+            <auro-alaska alaskaTagline></auro-alaska>
           </div>
           <div>
-            <auro-alaska style="width: 72px" customSize alaska></auro-alaska><br>
-            <auro-alaska style="width: 72px" customSize alaskaTagline></auro-alaska>
+            <auro-alaska style="width: 72px" alaska></auro-alaska><br>
+            <auro-alaska style="width: 72px" alaskaTagline></auro-alaska>
           </div>
         </template>
       </demo-snippet>
@@ -48,20 +48,20 @@
       <demo-snippet>
         <template>
           <div style="width: 528px" class="wrapper">
-            <auro-alaska customSize alaska ondark></auro-alaska>
-            <auro-alaska customSize alaskaTagline ondark></auro-alaska>
+            <auro-alaska alaska ondark></auro-alaska>
+            <auro-alaska alaskaTagline ondark></auro-alaska>
           </div>
           <div style="width: 192px" class="wrapper">
-            <auro-alaska customSize alaska ondark></auro-alaska>
-            <auro-alaska customSize alaskaTagline ondark></auro-alaska>
+            <auro-alaska alaska ondark></auro-alaska>
+            <auro-alaska alaskaTagline ondark></auro-alaska>
           </div>
           <div style="width: 108px" class="wrapper">
-            <auro-alaska customSize alaska ondark></auro-alaska>
-            <auro-alaska customSize alaskaTagline ondark></auro-alaska>
+            <auro-alaska alaska ondark></auro-alaska>
+            <auro-alaska alaskaTagline ondark></auro-alaska>
           </div>
           <div class="wrapper">
-            <auro-alaska style="width: 72px" customSize alaska ondark></auro-alaska><br>
-            <auro-alaska style="width: 72px" customSize alaskaTagline ondark></auro-alaska>
+            <auro-alaska style="width: 72px" alaska ondark></auro-alaska><br>
+            <auro-alaska style="width: 72px" alaskaTagline ondark></auro-alaska>
           </div>
         </template>
       </demo-snippet>
