+++ src/auro-alaska.js
@@ -19,7 +19,7 @@ import asTag100 from '@alaskaairux/icons/dist/restricted/AS-tagline-100_es6.js';
 
 // Import touch detection lib
 // import "focus-visible/dist/focus-visible.min.js";
-import styleCss from "./style-css.js";
+import styleCss from "./alaskaStyle-css.js";
 
 // See https://git.io/JJ6SJ for "How to document your components using JSDoc"
 /**
@@ -117,9 +117,10 @@ class AuroAlaska extends BaseIcon {
   }
 
   static get styles() {
-    return css`
-      ${styleCss}
-    `;
+    return [
+      super.styles,
+      css`${styleCss}`
+    ];
   }
 
   // function that renders the HTML and CSS into  the scope of the component
